Fix asset count
